Move dataset module to classes
